add dateUtils.js and update DeletingEntryModal.jsx, TodayWaterList.jsx, TodayListModal.jsx,
