+++ src/components/Home/DeletingEntryModal/DeletingEntryModal.jsx
@@ -1,3 +1,5 @@
+import { useDispatch } from 'react-redux';
+import { deleteWaterThunk } from '../../../redux/waterData/waterOperations';
 import { BaseModalWindow } from 'components';
 import {
   ButtonStyle,
@@ -6,13 +8,19 @@ import {
   ButtonBox,
 } from './DeletingEntryModal.styled';
 
-export const DeletingEntryModal = ({ onClose, onDelete, title }) => {
+export const DeletingEntryModal = ({ onClose, recordId }) => {
+  const dispatch = useDispatch();
+
+  const handleDelete = () => {
+    dispatch(deleteWaterThunk(recordId)).unwrap();
+    onClose();
+  };
   return (
-    <BaseModalWindow onClose={onClose} title={title}>
+    <BaseModalWindow onClose={onClose} title={'Delete Entry'}>
       <ModalContainer>
         <TextStyle>Are you sure you want to delete the entry?</TextStyle>
         <ButtonBox>
-          <ButtonStyle onClick={onDelete}>Delete</ButtonStyle>
+          <ButtonStyle onClick={handleDelete}>Delete</ButtonStyle>
           <ButtonStyle onClick={onClose}>Cancel</ButtonStyle>
         </ButtonBox>
       </ModalContainer>
