+++ src/helpers/utils/dateUtils.js
@@ -1,0 +1,34 @@
+export function formatTime(isoDate) {
+  return new Date(isoDate).toLocaleTimeString([], {
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: true,
+  });
+}
+
+export function formatTimeForInput(isoDate) {
+  return new Date(isoDate).toLocaleTimeString('en-GB', {
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: false,
+  });
+}
+
+export const formatMonth = monthYearString => {
+  const [year, month] = monthYearString.split('-').map(Number);
+  const monthNames = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December',
+  ];
+  return monthNames[month - 1];
+};
